+++ examples/hellostackwrightnext/CHANGELOG.md
@@ -1,5 +1,13 @@
 # stackwright-example-app
 
+## 0.2.1-alpha.1
+
+### Patch Changes
+
+- Updated dependencies [a5b331f]
+  - @stackwright/core@0.7.0-alpha.1
+  - @stackwright/nextjs@0.3.1-alpha.1
+
 ## 0.2.1-alpha.0
 
 ### Patch Changes